Change: Move get_last_release_version to the pontos.version module
